chore(automation):  + release and package the ingestion automation docker to official  + version 1.0.0  + e2e tests  + watch & manual watch  + compatible to InNetA

+++ server_automation/tests/test_ingestion_discrete.py
@@ -226,28 +226,30 @@ def teardown_module(module):  # pylint: disable=unused-argument
     This method been executed after test running - env cleaning
     """
     stop_watch()
-    if config.VALIDATION_SWITCH:
-        if config.TEST_ENV == config.EnvironmentTypes.QA.name or config.TEST_ENV == config.EnvironmentTypes.DEV.name:
-            # ToDo : Handle PVC - test it
-            try:
-                resp = azure_pvc_api.delete_ingestion_directory(api=config.PVC_DELETE_DIR)
-            except Exception as e:
-                resp = None
-                error_msg = str(e)
-            assert resp, \
-                f'Test: [{test_watch_discrete_ingest.__name__}] Failed: on following ingestion process (Folder delete) :  [{error_msg}]'
-            _log.info(f'Teardown - Finish PVC folder deletion')
+    if config.CLEAN_UP:
+        if config.VALIDATION_SWITCH:
+            if config.TEST_ENV == config.EnvironmentTypes.QA.name or config.TEST_ENV == config.EnvironmentTypes.DEV.name:
+                # ToDo : Handle PVC - test it
+                try:
+                    error_msg = None
+                    resp = azure_pvc_api.delete_ingestion_directory(api=config.PVC_DELETE_DIR)
+                except Exception as e:
+                    resp = None
+                    error_msg = str(e)
+                assert resp, \
+                    f'Test: [{test_watch_discrete_ingest.__name__}] Failed: on following ingestion process (Folder delete) :  [{error_msg}]'
+                _log.info(f'Teardown - Finish PVC folder deletion')
 
-        elif config.TEST_ENV == config.EnvironmentTypes.PROD.name:
-            if os.path.exists(config.NFS_ROOT_DIR_DEST):
-                shutil.rmtree(config.NFS_ROOT_DIR_DEST)
-                _log.info(f'Teardown - Finish NFS folder deletion')
+            elif config.TEST_ENV == config.EnvironmentTypes.PROD.name:
+                if os.path.exists(config.NFS_ROOT_DIR_DEST):
+                    shutil.rmtree(config.NFS_ROOT_DIR_DEST)
+                    _log.info(f'Teardown - Finish NFS folder deletion')
 
+                else:
+                    raise NotADirectoryError(
+                        f'Failed to delete directory because it doesnt exists: [{config.NFS_ROOT_DIR_DEST}]')
             else:
-                raise NotADirectoryError(
-                    f'Failed to delete directory because it doesnt exists: [{config.NFS_ROOT_DIR_DEST}]')
-        else:
-            raise ValueError(f'Illegal environment value type: {config.TEST_ENV}')
+                raise ValueError(f'Illegal environment value type: {config.TEST_ENV}')
     if config.CLEAN_UP and config.DEBUG_MODE_LOCAL:
         for p in ValueStorage.discrete_list:
             cleanup_env(p['product_id'], p['product_version'], initial_mapproxy_config)
